+++ weblate/settings_example.py
@@ -378,11 +378,17 @@ REGISTRATION_ALLOW_DISPOSABLE_EMAILS = False
 # Restrict private webhook targets
 # WEBHOOK_RESTRICT_PRIVATE = True
 
+# Restrict private asset targets
+# ASSET_RESTRICT_PRIVATE = True
+
 # Restrict private VCS repository targets
 # VCS_RESTRICT_PRIVATE = True
 
 # Private webhook target allowlist
 # WEBHOOK_PRIVATE_ALLOWLIST = [".internal.example", "hooks.internal.example"]
+
+# Private asset target allowlist
+# ASSET_PRIVATE_ALLOWLIST = [".internal.example", "assets.internal.example"]
 
 # Shortcut for login required setting
 REQUIRE_LOGIN = False